requirements.txt
pandas
pymongo
python-dotenv
sentence-transformers
tqdm